solving merge conflicts with UP and UU

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/model/ProjectRepository.java
@@ -14,5 +14,5 @@ public interface ProjectRepository extends CrudRepository<Project, Integer> {
     // you can join multiple attributes with and, e.g. findByProjectNameAndId for attributes ProjectName, Id
     List<Project> findByProjectName(String projectName);
 
-    Project findById(int id);
+    Project findById(Long id);
 }
